refactor: moved filter fragment classes into a package
fix: home fragment was making duplicate DateFilters
