+++ app/src/main/java/com/example/househomey/Filters/KeywordFilterFragment.java
@@ -1,4 +1,4 @@
-package com.example.househomey;
+package com.example.househomey.Filters;
 
 import android.view.View;
