+++ app/src/main/java/com/example/househomey/Filters/TagFilterFragment.java
@@ -1,10 +1,10 @@
-package com.example.househomey;
+package com.example.househomey.Filters;
 
 import android.view.View;
 
-public class MakeFilterFragment extends FilterFragment {
+public class TagFilterFragment extends FilterFragment {
 
-    public MakeFilterFragment(String title, View contentView) {
+    public TagFilterFragment(String title, View contentView) {
         super(title, contentView);
     }
